test(raijin): colocate initializer boundary tests
